refactor(interactive-chart): move mockup config to new file
